GH-40154: [C++][Parquet] Separate encoders and decoder
